BCPy2000 updated corresponding to third build release (MPI svn 13582)
git-svn-id: http://www.bci2000.org/svn/trunk@2546 7ef59cdf-0624-0410-a474-e34ce222aafa

+++ src/contrib/BCPy2000/framework/BCPy2000/BCI2000Tools/AudioSourceModule.py
@@ -90,8 +90,8 @@ class BciSource(BciGenericSource):
 		x = numpy.asarray(x[:nch, :])
 		out = numpy.asarray(sig[:nch, :])
 		#sig[nch:, :] = 0
-		packetsize = self.nominal['SamplesPerPacket']
-		sig[nch:, :] = self.packet_count * packetsize + numpy.array(range(packetsize), ndmin=2).repeat(sig.shape[0]-nch, axis=0)
+		packetsize = int(self.nominal['SamplesPerPacket'])
+		sig[nch:, :] = self.packet_count * packetsize + numpy.array(range(packetsize), ndmin=2, dtype='float').repeat(sig.shape[0]-nch, axis=0)
 
 		if self.use_env:
 			for i in xrange(sig.shape[1]):
